+++ test/System/Process/TypedSpec.hs
@@ -4,17 +4,15 @@ module System.Process.TypedSpec (spec) where
 
 import System.Process.Typed
 import System.IO
-import Data.Conduit
-import qualified Data.Conduit.Binary as CB
-import Network.HTTP.Simple
 import Control.Concurrent.Async (Concurrently (..))
+import Control.Concurrent.STM (atomically)
 import Test.Hspec
 import System.Exit
 import System.IO.Temp
 import qualified Data.ByteString as S
+import qualified Data.ByteString.Lazy as L
 import Data.String (IsString)
 import Data.Monoid ((<>))
-import qualified Data.Conduit.List as CL
 import qualified Data.ByteString.Base64 as B64
 
 #if !MIN_VERSION_base(4, 8, 0)
@@ -76,16 +74,15 @@ spec = do
         runProcess_ "false" `shouldThrow` \ExitCodeException{} -> True
 
     it "async" $ withSystemTempFile "httpbin" $ \fp h -> do
-        bss <- withProcess (setStdin createSink $ setStdout createSource "base64") $ \p ->
+        lbs <- withProcess (setStdin createPipe $ setStdout byteStringOutput "base64") $ \p ->
             runConcurrently $
-                Concurrently
-                    ( httpSink "https://raw.githubusercontent.com/fpco/typed-process/master/README.md" $ \_res ->
-                    CB.conduitHandle h .| getStdin p) *>
-                Concurrently
-                    ( runConduit
-                    $ getStdout p
-                   .| CL.consume)
+                Concurrently (do
+                  bs <- S.readFile "README.md"
+                  S.hPut h bs
+                  S.hPut (getStdin p) bs
+                  hClose (getStdin p)) *>
+                Concurrently (atomically $ getStdout p)
         hClose h
-        let encoded = S.filter (/= 10) $ S.concat bss
+        let encoded = S.filter (/= 10) $ L.toStrict lbs
         raw <- S.readFile fp
         encoded `shouldBe` B64.encode raw